fix some bug

+++ src/pages/products/list.jsx
@@ -1,4 +1,4 @@
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import Box from '@mui/material/Box';
 import Paper from '@mui/material/Paper';
 import Table from '@mui/material/Table';
@@ -16,11 +16,6 @@ import {
   CircularProgress,
   Alert,
   Collapse,
-  Dialog,
-  DialogTitle,
-  DialogContent,
-  DialogActions,
-  Button,
   Chip,
   Stack,
   Tooltip,
@@ -36,6 +31,9 @@ import { DashboardContent } from '../../layouts/dashboard/main';
 import { AdminPageHeader } from '../../components/AdminPageHeader';
 import { Helmet } from 'react-helmet-async';
 import { CONFIG } from '../../config-global';
+import useFetchProducts from '../../hooks/apis/useFetchProducts';
+import ConfirmDeleteModal from '../../components/ModalConfirmDelete';
+import { toast } from 'react-toastify';
 
 function Row(props) {
   const navigate = useNavigate();
@@ -92,9 +90,7 @@ function Row(props) {
               </IconButton>
             </Tooltip>
             <Tooltip title="Detail">
-              <IconButton
-                onClick={() => navigate(`/product/detail/${row.id}`)}
-              >
+              <IconButton onClick={() => navigate(`/product/detail/${row.id}`)}>
                 <InfoIcon />
               </IconButton>
             </Tooltip>
@@ -148,14 +144,10 @@ function Row(props) {
 
 const ProductsListAdmin = () => {
   const navigate = useNavigate();
-  const [products, setProducts] = useState([]);
   const [page, setPage] = useState(0);
   const [rowsPerPage, setRowsPerPage] = useState(10);
-  const [totalProducts, setTotalProducts] = useState(0);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState(null);
-  const [openDialog, setOpenDialog] = useState(false);
-  const [productToDelete, setProductToDelete] = useState(null);
+  const [isModalOpenDel, setModalOpenDel] = useState(false);
+  const [itemDel, setItemDel] = useState(null);
   const [filters, setFilters] = useState({
     search: '',
     category: [],
@@ -167,34 +159,8 @@ const ProductsListAdmin = () => {
     stockValue: 0,
   });
 
-  useEffect(() => {
-    const fetchProducts = async () => {
-      setLoading(true);
-      setError(null);
-      try {
-        const response = await axiosClient.get(
-          `${API_ROOT}/admin/product/list`,
-          {
-            params: {
-              page: page + 1,
-              limit: rowsPerPage,
-              ...filters,
-            },
-          },
-        );
-        setProducts(response.data.products);
-        setTotalProducts(response.data.totalProducts);
-        setPage(0);
-      } catch (error) {
-        setError('Error fetching products. Please try again later.');
-        console.error('Error fetching products:', error);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchProducts();
-  }, [page, rowsPerPage, filters]);
+  const { products, totalProducts, loading, error, fetchProducts } =
+    useFetchProducts(page, rowsPerPage, filters);
 
   const handleChangePage = (event, newPage) => {
     setPage(newPage);
@@ -205,41 +171,22 @@ const ProductsListAdmin = () => {
     setPage(0);
   };
 
-  const handleOpenDialog = productId => {
-    setProductToDelete(productId);
-    setOpenDialog(true);
+  const handleOpenModalDel = id => {
+    setItemDel(id);
+    setModalOpenDel(true);
   };
+  const handleCloseModalDel = () => setModalOpenDel(false);
 
-  const handleCloseDialog = () => {
-    setOpenDialog(false);
-    setProductToDelete(null);
-  };
-
-  const handleConfirmDelete = async () => {
-    if (productToDelete) {
-      try {
-        await axiosClient.delete(
-          `${API_ROOT}/admin/product/delete/${productToDelete}`,
-        );
-
-        const response = await axiosClient.get(
-          `${API_ROOT}/admin/product/list`,
-          {
-            params: {
-              page: page + 1,
-              limit: rowsPerPage,
-            },
-          },
-        );
-        setProducts(response.data.products);
-        setTotalProducts(response.data.totalProducts);
-
-        handleCloseDialog();
-      } catch (error) {
-        setError('Error deleting product. Please try again later.');
-        console.error('Error deleting product', error);
-      }
+  const handleConfirmDel = async () => {
+    try {
+      await axiosClient.delete(`${API_ROOT}/admin/product/delete/${itemDel}`);
+      setItemDel(null);
+      fetchProducts();
+      toast.success('Deleted product succsessfully!');
+    } catch (error) {
+      toast.error('Delete product failed!');
     }
+    handleCloseModalDel();
   };
 
   return (
@@ -316,7 +263,7 @@ const ProductsListAdmin = () => {
                         <Row
                           key={row.id}
                           row={row}
-                          handleOpenDialog={handleOpenDialog}
+                          handleOpenDialog={handleOpenModalDel}
                         />
                       ))}
                     </TableBody>
@@ -336,28 +283,11 @@ const ProductsListAdmin = () => {
           </>
         )}
 
-        <Dialog
-          open={openDialog}
-          onClose={handleCloseDialog}
-          aria-labelledby="alert-dialog-title"
-          aria-describedby="alert-dialog-description"
-        >
-          <DialogTitle id="alert-dialog-title">Confirm Deletion</DialogTitle>
-          <DialogContent>
-            <Typography variant="body1" color="textSecondary">
-              Are you sure you want to delete this product? This action cannot
-              be undone.
-            </Typography>
-          </DialogContent>
-          <DialogActions>
-            <Button onClick={handleCloseDialog} color="secondary">
-              Cancel
-            </Button>
-            <Button onClick={handleConfirmDelete} color="primary" autoFocus>
-              Delete
-            </Button>
-          </DialogActions>
-        </Dialog>
+        <ConfirmDeleteModal
+          open={isModalOpenDel}
+          onConfirm={handleConfirmDel}
+          onCancel={handleCloseModalDel}
+        />
       </DashboardContent>
     </>
   );
